Refactor : 멘토 리뷰 반영 / footer 상수 데이터로 변경
